+++ mobile/src/pages/OrphanageDetails.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import { Text } from "react-native"
+
+const OrphanagesDetails = () => {
+  return <Text>Details</Text>
+}
+
+export default OrphanagesDetails
