+++ mobile/src/pages/OrphanagesMap.tsx
@@ -1,0 +1,144 @@
+import React from "react"
+import {
+  Dimensions,
+  SafeAreaView,
+  StatusBar,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+} from "react-native"
+import MapView, { Callout, Marker, PROVIDER_GOOGLE } from "react-native-maps"
+import { Feather } from "@expo/vector-icons"
+import { useFonts } from "expo-font"
+import {
+  Nunito_600SemiBold,
+  Nunito_700Bold,
+  Nunito_800ExtraBold,
+} from "@expo-google-fonts/nunito"
+
+import mapMarker from "../images/map-marker.png"
+import { useNavigation } from "@react-navigation/native"
+
+const OrphanagesMap = () => {
+  const navigation = useNavigation()
+  const [fontsLoaded] = useFonts({
+    Nunito_600SemiBold,
+    Nunito_700Bold,
+    Nunito_800ExtraBold,
+  })
+
+  if (!fontsLoaded) {
+    return null
+  }
+
+  function handleNavigateToOrphanageDetail() {
+    navigation.navigate("OrphanageDetails")
+  }
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <StatusBar barStyle="light-content" backgroundColor="#0089a5" />
+      <MapView
+        provider={PROVIDER_GOOGLE}
+        style={styles.map}
+        initialRegion={{
+          latitude: -4.0318947,
+          longitude: -38.3060312,
+          latitudeDelta: 0.008,
+          longitudeDelta: 0.008,
+        }}
+      >
+        <Marker
+          icon={mapMarker}
+          calloutAnchor={{
+            x: 2.7,
+            y: 0.8,
+          }}
+          coordinate={{
+            latitude: -4.0318947,
+            longitude: -38.3060312,
+          }}
+        >
+          <Callout tooltip onPress={handleNavigateToOrphanageDetail}>
+            <View style={styles.calloutContainer}>
+              <Text style={styles.calloutText}>Lar das meninas</Text>
+            </View>
+          </Callout>
+        </Marker>
+      </MapView>
+
+      <View style={styles.footer}>
+        <Text style={styles.footerText}>2 Orfanatos encontrado</Text>
+        <TouchableOpacity
+          style={styles.createOrphanageButton}
+          onPress={() => {}}
+        >
+          <Feather name="plus" size={20} color="#fff" />
+        </TouchableOpacity>
+      </View>
+    </SafeAreaView>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    height: Dimensions.get("window").height,
+  },
+
+  map: {
+    width: Dimensions.get("window").width,
+    height: Dimensions.get("window").height,
+  },
+
+  calloutContainer: {
+    width: 160,
+    height: 46,
+    paddingHorizontal: 16,
+    backgroundColor: "rgba(255, 255, 255, 0.8)",
+    borderRadius: 16,
+    justifyContent: "center",
+  },
+
+  calloutText: {
+    color: "#0089a5",
+    fontSize: 16,
+    fontFamily: "Nunito_700Bold",
+  },
+
+  footer: {
+    position: "absolute",
+    left: 24,
+    right: 24,
+    bottom: 32,
+
+    backgroundColor: "#fff",
+    borderRadius: 15,
+    height: 46,
+    paddingLeft: 24,
+
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+
+    elevation: 4,
+  },
+
+  footerText: {
+    color: "#8fa7b3",
+    fontFamily: "Nunito_700Bold",
+  },
+
+  createOrphanageButton: {
+    width: 46,
+    height: 46,
+    backgroundColor: "#15c3d6",
+    borderRadius: 15,
+
+    justifyContent: "center",
+    alignItems: "center",
+  },
+})
+
+export default OrphanagesMap
